Combine code for road proxy generation. This aligned code that was originally in the "sandbox" subdir on this branch
