+++ gch4i/proxy_processing/task_roads_proxy.py
@@ -1,4 +1,19 @@
-########################################################################################
+'''
+THE FINAL OUTPUT FOR THIS SCRIPT CAN BE FOUND ON ONE DRIVE/SHARE POINT AT THE FOLLOWING PATH:
+C:\Users\<USERNAME>\Environmental Protection Agency (EPA)\Gridded CH4 Inventory - RTI 2024 Task Order\Task 2\ghgi_v3_working\v3_data\global\raw_roads\task_outputs\gridwise_outputs\*cell.csv
+
+The files named gridwise_mapping_YYYY_cell.csv are the estimated methane output aggregated to the cell level.
+    The variable for methane emission allocation ('methane_emission_allocation') represents the proportion of 
+    state-wide methane emissions from that year that are attirbutable to that vehicle type/road/urbanicity(region) type in a given cell
+
+The files named gridwise_mapping_YYYY.csv are the estimated methane ouput for each cell/state/region overlay by vehicle type and road type.
+    This file is the fully disaggregated methane emission. The variable for methane emission in this dataset represents more of a "rate" that needs to be divided by m_road_in_state to achieve true estimate of methane emission ((long_table['rd_length'] / long_table['m_road_in_state']) * long_table['proxy'])
+
+The files named gridwise_mapping_intermediate_YYYY.csv is intermediate output that is used to calculate the final output.
+    This file is similar to the gridwise_mapping_YYYY.csv file, but it contains the raw proxy and road length for that overlay
+'''
+
+
 # %% STEP 0.1. Load Packages
 
 from pathlib import Path
@@ -6,12 +21,19 @@ import os
 from typing import Annotated
 from pytask import Product, mark, task
 
+from shapely import errors as se
+from shapely import from_wkt
+
 import pandas as pd
 import numpy as np
 import gc
 
+from datetime import datetime
+
 import geopandas as gpd
-from shapely.geometry import MultiLineString, LineString, GeometryCollection
+import dask_geopandas as dgpd
+from pyproj import CRS
+from shapely.geometry import MultiLineString, LineString, GeometryCollection, box
 
 from gch4i.config import (
     V3_DATA_PATH,
@@ -21,6 +43,9 @@ from gch4i.config import (
     years,
     load_state_ansi
 )
+
+from gch4i.gridding import GEPA_spatial_profile
+from gch4i.utils import get_cell_gdf
 
 """
 The global function path is commented out because it would work properly when running
@@ -73,9 +98,131 @@ num_years = len(year_range)
 # %% Functions
 ########################################################################################
 
+def get_overlay_dir(year, 
+                    out_dir: Path=task_outputs_path / 'overlay_cell_state_region'):
+    return out_dir / f'cell_state_region_{year}.parquet'
+
+def get_overlay_gdf(year, crs="EPSG:4326"):
+    crs_obj = CRS(crs)
+
+    gdf = gpd.read_parquet(get_overlay_dir(year))
+
+    if gdf.crs != crs_obj:
+        print(f"Converting overlay to crs {crs_obj.to_epsg()}")
+        gdf.to_crs(crs, inplace=True)
+
+    return gdf
+
+# Read in State Spatial Data
+def get_states_gdf(crs="EPSG:4326"):
+    """
+    Read in State spatial data
+    """
+    crs_obj = CRS(crs)
+
+    gdf_states = gpd.read_file(gdf_state_files)
+
+
+    gdf_states = gdf_states[~gdf_states['STUSPS'].isin(
+        ['VI', 'MP', 'GU', 'AS', 'PR', 'AK', 'HI']
+        )]
+    gdf_states = gdf_states[['STUSPS', 'NAME', 'geometry']]
+
+    if gdf_states.crs != crs_obj:
+        print(f"Converting states to crs {crs_obj.to_epsg()}")
+        gdf_states.to_crs(crs, inplace=True)
+
+    return gdf_states
+
+# Read in Region Spatial Data
+def get_region_gdf(year, crs=4326):
+    """
+    Read in region spatial data
+    """
+    crs = CRS(crs)
+    road_loc = (
+        gpd.read_parquet(f"{road_file}{year}_us_uac.parquet", columns=['geometry'])
+        .assign(year=year)
+        .to_crs(crs)
+        .assign(urban=1)
+    )
+    return road_loc
+
+def benchmark_load(func):
+    def wrapper(*args, **kwargs):
+        start = datetime.now()
+        result = func(*args, **kwargs)
+        print(f"{func.__name__} took {datetime.now() - start}")
+        return result
+    return wrapper
+
+def get_roads_path(year, raw_roads_path: Path=V3_DATA_PATH / "global/raw_roads", raw=True):
+    '''
+    If raw is True, it returns the parquet files of the original roads data at at path:
+        C:/Users/<USER>/Environmental Protection Agency (EPA)/Gridded CH4 Inventory - RTI 2024 Task Order/Task 2/ghgi_v3_working/v3_data/global/raw_roads/tl_{year}_us_allroads.parquet
+    If raw is False, it returns the parquet files of the reduced roads data (as generated in Andrew's original task_roads_proxy.py script `reduce_roads()`): 
+        C:/Users/<USER>/Environmental Protection Agency (EPA)/Gridded CH4 Inventory - RTI 2024 Task Order/Task 2/ghgi_v3_working/v3_data/global/raw_roads/task_outputs/reduced_roads_{year}.parquet
+    
+    '''
+    if raw:
+        return Path(raw_roads_path) / f"tl_{year}_us_allroads.parquet"
+    else:
+        return Path(raw_roads_path) / f"task_outputs/reduced_roads_{year}.parquet"
+
+
+@benchmark_load
+def read_roads(year, raw=True, crs='EPSG:4326'):
+    crs_obj = CRS(crs)
+    gdf = gpd.read_parquet(get_roads_path(year, raw=raw))
+    if gdf.crs != crs_obj:
+        print(f"Converting {year} roads to crs {crs_obj.to_epsg()}")
+        return gdf.to_crs(crs)
+    else:
+        return gdf
+    
+def intersect_sindex(cell, roads):
+    '''
+    Based of fthis geoff boeing blog post:
+    https://geoffboeing.com/2016/10/r-tree-spatial-index-python/
+    '''
+    # first, add rtree spatial index to roads
+
+    
+    spatial_index = roads.sindex
+    possible_matches_index = list(spatial_index.intersection(cell.bounds))
+    possible_matches = roads.iloc[possible_matches_index]
+    precise_matches = possible_matches[possible_matches.intersects(cell)]
+    return precise_matches
+
+# @benchmark_load
+def intersect_and_clip(roads, state_geom, state_abbrev=None, simplify=None, dask=False):
+    '''
+    The simplify parameer is used to simplify the geometry of the roads before clipping to the state boundary
+    '''
+    state_box = box(*state_geom.bounds)
+    # Get the roads that intersect with the state bounding box
+    state_roads = intersect_sindex(state_box, roads)
+    # print(f"Found {len(state_roads)} roads in {state_abbrev}")
+    if not state_roads.empty:
+        if simplify:
+            state_geom = state_geom.simplify(simplify)
+        
+        # convert to dask
+        if dask:
+            state_roads = dgpd.from_geopandas(state_roads, npartitions=4)
+            # clip roads to state boundary
+            result = state_roads.clip(state_geom)
+            return result.compute()
+        
+        else:
+            # clip roads to state boundary
+            result = state_roads.clip(state_geom)
+            return result
+    return state_roads
+
 
 # Read in VM2 Data
-def read_vmt2(num_years):
+def get_vm2_arrays(num_years):
     # Initialize arrays
     Miles_road_primary = np.zeros([2, len(State_ANSI), num_years])
     Miles_road_secondary = np.zeros([2, len(State_ANSI), num_years])
@@ -140,7 +287,7 @@ def read_vmt2(num_years):
 
 
 # Read in VM4 Data
-def read_vmt4(num_years):
+def get_vm4_arrays(num_years):
     # Initialize arrays
     Per_vmt_mot = np.zeros([2, 3, len(State_ANSI), num_years])
     Per_vmt_pas = np.zeros([2, 3, len(State_ANSI), num_years])
@@ -406,6 +553,7 @@ def calculate_state_proxies(num_years,
             tot_hea[istate, iyear] = np.sum(vmt_hea[:, :, istate, iyear])
 
     # Calculate proxy values
+    print(f"CALCULATING")
     pas_proxy = vmt_pas / tot_pas
     lig_proxy = vmt_lig / tot_lig
     hea_proxy = vmt_hea / tot_hea
@@ -414,7 +562,7 @@ def calculate_state_proxies(num_years,
 
 
 # Unpack State Proxy Arrays
-def unpack_state_proxy(state_proxy_array):
+def unpack_state_proxy(state_proxy_array, proxy_name='Emission Allocation'):
     reshaped_state_proxy = state_proxy_array.reshape(-1)
 
     row_index = np.repeat(['urban', 'rural'], 3 * 57 * num_years)
@@ -427,7 +575,7 @@ def unpack_state_proxy(state_proxy_array):
         'Road Type': col1_index,
         'State': col2_index,
         'Year': col3_index,
-        'Proxy': reshaped_state_proxy
+        proxy_name: reshaped_state_proxy
     })
 
     df['State_abbr'] = df['State'].map(state_mapping)
@@ -466,10 +614,11 @@ def unpack_state_allroads_proxy(vmt_tot):
 
 
 # Generate Roads Proportions Data
-def get_roads_proportion_data(pas_proxy, lig_proxy, hea_proxy):
+def get_roads_proportion_data(pas_proxy, lig_proxy, hea_proxy, out_path, proxy_name='Emission Allocation'):
     """
     Formats data for roads proxy emissions
     """
+    proxy_name = proxy_name.lower()
     # Add Vehicle Type column
     pas_proxy['Vehicle'] = 'Passenger'
     lig_proxy['Vehicle'] = 'Light'
@@ -486,9 +635,9 @@ def get_roads_proportion_data(pas_proxy, lig_proxy, hea_proxy):
                        .query("state_code not in ['VI', 'MP', 'GU', 'AS', 'PR', 'AK', 'HI', 'UM']")
     )
     vmt_roads_proxy = vmt_roads_proxy[['state_code', 'year', 'vehicle', 'region',
-                                       'road_type', 'proxy']]
+                                       'road_type', proxy_name]]
 
-    vmt_roads_proxy.to_csv(task_outputs_path / "roads_proportion_data.csv", index=False)
+    vmt_roads_proxy.to_csv(out_path, index=False)
 
     del pas_proxy, lig_proxy, hea_proxy
 
@@ -497,708 +646,240 @@ def get_roads_proportion_data(pas_proxy, lig_proxy, hea_proxy):
     return None
 
 
-########################################################################################
-# %% Proxy Functions
+def get_road_proxy_data(road_proxy_out_path: Path=task_outputs_path / "roads_proportion_data.csv"):
+    if not road_proxy_out_path.exists():
+        # Proportional Allocation of Roads Emissions
+        # The "Proxy" column in the output data represents the proportion of VMT for each vehicle type, disaggregated by urban/rural and road type
+        #################
+        # VM2 Outputs
+        Miles_road_primary, Miles_road_secondary, Miles_road_other, total, total2 = get_vm2_arrays(num_years)
 
-def read_reduce_data(year):
-    """
-    Read in All Roads data. Deduplicate and reduce data early.
-    """
-    # Order road types
-    road_type_order = ['Primary', 'Secondary', 'Other']
+        # VM4 Outputs
+        Per_vmt_mot, Per_vmt_pas, Per_vmt_lig, Per_vmt_hea = get_vm4_arrays(num_years)
 
-    df = (gpd.read_parquet(f"{raw_road_file}{year}_us_allroads.parquet",
-                           columns=['MTFCC', 'geometry'])
-          .assign(year=year))
+        # State Proxy Outputs
+        pas_proxy, lig_proxy, hea_proxy, vmt_tot = calculate_state_proxies(num_years,
+                                                                        Miles_road_primary,
+                                                                        Miles_road_secondary,
+                                                                        Miles_road_other,
+                                                                        Per_vmt_pas,
+                                                                        Per_vmt_lig,
+                                                                        Per_vmt_hea)
 
-    road_data = (
-        df.to_crs("ESRI:102003")
-        .assign(
-            geometry=lambda df: df.normalize(),
-            road_type=lambda df: pd.Categorical(
-                np.select(
-                    [
-                        df['MTFCC'] == 'S1100',
-                        df['MTFCC'] == 'S1200',
-                        df['MTFCC'].isin(['S1400', 'S1630', 'S1640'])
-                    ],
-                    [
-                        'Primary',
-                        'Secondary',
-                        'Other'
-                    ],
-                    default=None
-                ),
-                categories=road_type_order,  # Define the categories
-                ordered=True  # Ensure the categories are ordered
-            )
-        )
-    )
-    # Sort
-    road_data = road_data.sort_values('road_type').reset_index(drop=True)
-    # Explode to make LineStrings
-    road_data = road_data.explode(index_parts=True).reset_index(drop=True)
-    # Remove duplicates of geometries
-    road_data = road_data.drop_duplicates(subset='geometry', keep='first')
+        # Unpack State Proxy Outputs
+        pas_proxy = unpack_state_proxy(pas_proxy)
+        lig_proxy = unpack_state_proxy(lig_proxy)
+        hea_proxy = unpack_state_proxy(hea_proxy)
+        # tot_proxy = unpack_state_allroads_proxy(vmt_tot)
 
-    # Separate out Road Types
-    prim_year = road_data[road_data['road_type'] == 'Primary']
-    sec_year = road_data[road_data['road_type'] == 'Secondary']
-    oth_year = road_data[road_data['road_type'] == 'Other']
+        # Generate Roads Proportions Data
+        get_roads_proportion_data(pas_proxy, lig_proxy, hea_proxy, out_path=road_proxy_out_path)
 
-    buffer_distance = 3  # meters
-
-    # Set buffers
-    prim_buffer = prim_year.buffer(buffer_distance)
-    prim_buffer = gpd.GeoDataFrame(geometry=prim_buffer, crs=road_data.crs)
-
-    prisec_buffer = pd.concat([prim_year, sec_year], ignore_index=True)
-    prisec_buffer = prisec_buffer.buffer(buffer_distance)
-    prisec_buffer = gpd.GeoDataFrame(geometry=prisec_buffer, crs=road_data.crs)
-
-    # Overlay
-    sec_red = gpd.overlay(sec_year, prim_buffer, how='difference')
-    other_red = gpd.overlay(oth_year, prisec_buffer, how='difference')
-
-    # Combine
-    road_data = pd.concat([prim_year, sec_red, other_red], ignore_index=True)
-
-    road_data = road_data[['year', 'road_type', 'geometry']]
-
-    # Write to parquet
-    road_data.to_parquet(task_outputs_path / f"reduced_roads_{year}.parquet")
-
-    del road_data, prim_year, sec_year, oth_year, prim_buffer, prisec_buffer, sec_red, other_red
-
-    gc.collect()
-
-    return None
+        return pas_proxy, lig_proxy, hea_proxy
+    
+    return pd.read_csv(road_proxy_out_path)
 
 
-# Read in Region Spatial Data
-def read_regions(year):
-    """
-    Read in region spatial data
-    """
-    road_loc = (
-        gpd.read_parquet(f"{road_file}{year}_us_uac.parquet", columns=['geometry'])
-        .assign(year=year)
-        .to_crs("ESRI:102003")
-    )
-    return road_loc
+@benchmark_load
+def overlay_cell_state_region(cell_gdf, region_gdf, state_gdf):
+    '''
+    This function overlays the cell grid with the state and region boundaries for each year of region data
+    '''    
+    # Overlay the cell grid with the state and region boundaries
+    print(f"Overlaying cell grid with state and region boundaries: {datetime.now()}")
+    cell_state_region_gdf = gpd.overlay(cell_gdf, state_gdf, how='union')
+    print(f"Overlayed cell grid with state boundaries: {datetime.now()}")
+    cell_state_region_gdf = gpd.overlay(cell_state_region_gdf, region_gdf, how='union')
+    
+    # where urban is NaN, set to 0 (since the "region" dataset is urban geometries) and drop variable for year (since it is redundant and in the file name)
+    cell_state_region_gdf['urban'] = cell_state_region_gdf['urban'].fillna(0).astype(int)
+    cell_state_region_gdf.drop(columns=['year'], inplace=True)
 
+    # drop rows with no cell_id, as this indicates that the geometry falls outside of the US and so doesn't need to be processed
+    cell_state_region_gdf.dropna(subset=['cell_id'], inplace=True)
 
-# Read in State Spatial Data
-def read_states():
-    """
-    Read in State spatial data
-    """
+    return cell_state_region_gdf
 
-    gdf_states = gpd.read_file(gdf_state_files)
-
-    gdf_states = gdf_states[~gdf_states['STUSPS'].isin(
-        ['VI', 'MP', 'GU', 'AS', 'PR', 'AK', 'HI']
-        )]
-    gdf_states = gdf_states[['STUSPS', 'NAME', 'geometry']]
-    gdf_states = gdf_states.to_crs("ESRI:102003")
-
-    return gdf_states
-
-
-# Extract Line geometries from Geometry Collections
-def extract_lines(geom):
-    if geom is None:
+@benchmark_load
+def run_overlay_for_year(year, out_dir):
+    # Save the overlaid geoparquet file
+    out_path = get_overlay_dir(year, out_dir)
+    if out_path.exists():
+        print(f"File already exists for {year}: {out_path}")
         return None
-    if isinstance(geom, (LineString, MultiLineString)):
-        # Return the geometry
-        return geom
-    elif isinstance(geom, GeometryCollection):
-        # Filter the collection to include only LineString or MultiLineString, Remove Points
-        lines = [g for g in geom.geoms if isinstance(g, (LineString, MultiLineString))]
-        if len(lines) == 1:
-            return lines[0]
-        elif len(lines) > 1:
-            return MultiLineString(lines)
+    
+    # Load the region, cell and state data
+    print(f"Reading datasets for {year}: {datetime.now()}")
+    cell_gdf = get_cell_gdf().to_crs(4326).reset_index().rename(columns={'index': 'cell_id'})
+    region_gdf = get_region_gdf(year)
+    state_gdf = get_states_gdf()
+
+    # Overlay the cell grid with the state and region boundaries
+    cell_state_region_gdf = overlay_cell_state_region(cell_gdf, region_gdf, state_gdf)
+
+    cell_state_region_gdf.to_parquet(out_path)
+
+    print(f"Saved overlaid geoparquet file for {year} to {out_dir / f'cell_state_region_{year}.parquet'}")
+
+
+@benchmark_load
+def overlay_roads_on_states(roads, states, dask=False):
+    if not dask:
+        # Overlay the roads with the state boundaries
+        print(f"Overlaying roads with state boundaries: {datetime.now()}")
+        roads = gpd.overlay(roads, states, how='intersection')
+        print(f"Overlay complete: {datetime.now()}")
+        return roads
     else:
-        return None  # Only returns LineStrings and MultiLineStrings (no polygons or points)
+        print(f'USING DASK')
+        # converting to dask:
+        # convert roads to a dask geopandas dataframe
+        roads = dgpd.from_geopandas(roads, npartitions=4)
+        # convert states to a dask geopandas dataframe
+        states = dgpd.from_geopandas(states, npartitions=4)
+
+        # Overlay the roads with the state boundaries
+        print(f"Overlaying roads with state boundaries: {datetime.now()}")
+        result = dgpd.overlay(roads, states, how='intersection')
+        print(f"Overlay complete: {datetime.now()}")
+        return result.compute()
+    
+
+#########################################################################
+#########################################################################
+# Main Script
+#########################################################################
+#########################################################################
 
 
-# Process geometry column
-def process_geometry_column(gdf):
-    # Apply the extract_lines function
-    gdf['geometry'] = gdf['geometry'].apply(extract_lines)
+#########################################################################
+# get gridded overlay and output to file
+out_dir = task_outputs_path / 'overlay_cell_state_region'
+out_dir.mkdir(parents=True, exist_ok=True)
 
-    # Remove rows where geometry is None
-    gdf = gdf.dropna(subset=['geometry'])
-
-    # Ensure the GeoDataFrame only contains LineString and MultiLineString
-    gdf = gdf[gdf['geometry'].apply(lambda geom: isinstance(geom, (LineString, MultiLineString)))]
-
-    # Explode MultiLineStrings into LineStrings
-    gdf = gdf.explode(index_parts=True).reset_index(drop=True)
-
-    return gdf
+print(f'RUNNING OVERLAY')
+print(f"NOTE: OVERLAY WILL TAKE ~2 HOURS PER YEAR TO RUN")
+for year in year_range:
+    run_overlay_for_year(year, out_dir)
 
 
-# Function to process each year's data
-def process_year(year_data, year_region):
-    # Ensure year_region only has necessary columns
-    year_region = year_region[['geometry']]
+#########################################################################
+# process the roads data
 
-    # Overlay for urban roads
-    urban_roads = gpd.overlay(year_data, year_region, how='intersection',
-                              keep_geom_type=False)
-    urban_roads['region'] = 'urban'
-
-    # Overlay for rural roads
-    rural_roads = gpd.overlay(year_data, year_region, how='difference',
-                              keep_geom_type=False)
-    rural_roads['region'] = 'rural'
-
-    # Combine results and select necessary columns
-    combined = pd.concat([urban_roads, rural_roads])
-
-    # Filter for LineStrings and MultiLineStrings
-    combined = process_geometry_column(combined)
-
-    return combined[['year', 'STUSPS', 'region', 'road_type', 'geometry']]
+urban_map = {1: 'urban', 0: 'rural'}
+task_outputs_path = task_outputs_path / "gridwise_outputs"
+task_outputs_path.mkdir(exist_ok=True, parents=True)
 
 
-# Finishing function
-def state_region_join(year):
+for year in range(min_year, max_year+1):
+        if year < 2018:
+            continue
+        epsg = "ESRI:102003"
+        road_type_col = 'road_type'
+        road_proxy_df = get_road_proxy_data().rename(columns={'state_code': 'state'})
+        cells = get_overlay_gdf(year, crs=epsg).rename(columns={'STUSPS': 'state', 'urban': 'region'})
+        cells = cells.loc[~cells['state'].isna()]
+        cells['region'] = cells['region'].map(urban_map)
 
-    states_gdf = read_states()
+        print(f'Reading roads')
+        roads = read_roads(year, crs=epsg, raw=False)
 
-    # Step 1: Filter to year
-    road_data = gpd.read_parquet(task_outputs_path / f"reduced_roads_{year}.parquet")
-    # Step 2: Overlay to cut at state boundaries
-    road_data = gpd.overlay(road_data, states_gdf, how='identity', keep_geom_type=False)
-    # Step 3: Remove geometries that re not LineStrings or MultiLineStrings               # Check here if an issue
-    road_data = process_geometry_column(road_data)
-    # Step 4: Use a state buffer to prevent accidental road reduction
-    buffer_distance = 10  # meters
-    states_gdf_buffer = states_gdf.buffer(buffer_distance)
-    states_gdf_buffer = gpd.GeoDataFrame(geometry=states_gdf_buffer, crs=states_gdf.crs)
-    states_gdf_buffer = states_gdf_buffer.join(states_gdf.drop(columns='geometry'))
-    # Step 5 Spatial join to assign state attributes
-    road_data = gpd.sjoin(road_data, states_gdf_buffer, how='left', predicate='within')
-    # Step 6: Remove duplicate road segments (again, if new generated)
-    road_data = road_data.drop_duplicates(subset=['geometry'])
-    # Step 7: Clean table
-    road_data = (
-        road_data[['year', 'STUSPS_right', 'geometry', 'road_type']]
-        .rename(columns={'STUSPS_right': 'STUSPS'})
-        .dropna(subset=['STUSPS'])
-        # .assign(year=year,
-        #         type=road_type_label)
-    )
-    # Step 8: Garbage collection
-    del states_gdf, states_gdf_buffer, buffer_distance
+        print(f"Starting processing for {year} at {datetime.now()}")
+        interm_out_path = task_outputs_path / f"gridwise_mapping_intermediate_{year}.csv"
 
-    gc.collect()
-
-    # Part 3: Generate Roads by State, Region
-    year_region = read_regions(year)
-    result = process_year(road_data, year_region)
-
-    # Combine results and dissolve
-    result = result.to_crs(epsg=4326)
-    result = result.dissolve(by=['year', 'STUSPS', 'region', 'road_type']).reset_index()
-
-    result.to_parquet(task_outputs_path / f"final_roads_{year}.parquet")
-
-    del road_data, year_region, result
-
-    gc.collect()
-
-    return None
-
-
-##########
-# %% Join Data Together
-# Proportional and Proxy Data Join
-def prop_proxy_join():
-    """
-    Join proportional and proxy data
-    """
-
-    # Read in Proportional Data
-    proportional_proxy = pd.read_csv(task_outputs_path / "roads_proxy_proportions.csv")
-
-    # Read in All Roads Data
-    result_list = []
-    for year in year_range:
-        year_data = gpd.read_parquet(task_outputs_path / f"final_roads_{year}.parquet")
-        result_list.append(year_data)
-    all_roads_proxy = pd.concat(result_list, ignore_index=True)
-
-    del result_list
-    gc.collect()
-
-    # Join proportional data with geometry data
-    roads_proxy = pd.merge(
-        proportional_proxy,
-        all_roads_proxy,
-        on=['state_code', 'year', 'region', 'road_type'],
-        how='left'
-    ).reset_index()
-
-    # Convert back to a GeoDataFrame
-    roads_proxy = gpd.GeoDataFrame(roads_proxy, geometry="geometry").to_crs(epsg=4326)
-
-    return roads_proxy
-
-
-########################################################################################
-########################################################################################
-########################################################################################
-# %% Pytask
-
-@mark.persist
-@task(id="roads_proxy")
-def task_get_roads_proxy(
-    state_path: Path = gdf_state_files,
-    reporting_proxy_output: Annotated[Path, Product] = proxy_data_dir_path
-    / "mobile_combustion/roads_proxy.parquet",
-):
-    """
-    Relative location information for roads in the US.
-    """
-
-    # Proportional Allocation of Roads Emissions
-    ####################################################################################
-    # VM2 Outputs
-    Miles_road_primary, Miles_road_secondary, Miles_road_other, total, total2 = read_vmt2(num_years)
-
-    # VM4 Outputs
-    Per_vmt_mot, Per_vmt_pas, Per_vmt_lig, Per_vmt_hea = read_vmt4(num_years)
-
-    # State Proxy Outputs
-    pas_proxy, lig_proxy, hea_proxy, vmt_tot = calculate_state_proxies(num_years,
-                                                                       Miles_road_primary,
-                                                                       Miles_road_secondary,
-                                                                       Miles_road_other,
-                                                                       Per_vmt_pas,
-                                                                       Per_vmt_lig,
-                                                                       Per_vmt_hea)
-
-    # Unpack State Proxy Outputs
-    pas_proxy = unpack_state_proxy(pas_proxy)
-    lig_proxy = unpack_state_proxy(lig_proxy)
-    hea_proxy = unpack_state_proxy(hea_proxy)
-    # tot_proxy = unpack_state_allroads_proxy(vmt_tot)
-
-    # Generate Roads Proportions Data
-    get_roads_proportion_data(pas_proxy, lig_proxy, hea_proxy)
-
-    # Roads Proxy Processing
-    ####################################################################################
-
-    # Read in All Roads data. Deduplicate and reduce data early.
-    for year in year_range:
-        read_reduce_data(year)
-
-    """
-    Current outputs ran through this point: read_reduce_data(year)
-    * output reduced_roads_{year}.parquet & roads_proportion_data.csv to raw_roads>task_outputs folder
-
-    Previous versions:
-    * read in all years to one gdf and processed by year
-        * can't hold all years in memory
-        * Solution: read in one year at a time to process and write to disk
-        * Doesn't need to be continuously read in and out, this was for testing steps
-
-    Attempted Solutions:
-    * DASK Geopandas : worked on one year at a time, took too long on region split
-    * Parallel Processing : Ran into issues with batching (Geopandas CRS)
-        * converted to WKT to batch and converted back to CRS when read in
-        * briefly worked, kernel died, stopped experimentation to focus on other tasks
-    * Sequential Processing (no parallel processing, just batched): 
-        * Built out code, did not run yet
-
-    File: processed_state_roads_2012.parquet
-    * stored in task_outputs folder
-    * took 2 hours to run
-    * effectively deduplicated roads and clipped to state boundaries
-        * Tested methods:
-            * grouping by state and clipping at state boundaries (running for each state)
-            was quicker than gpd.overlay( identity ) > add buffer to state polygons >
-            sjoin( within )
-    * Code was originally ran in a different script, will include below Pytask code
-
-    read_reduce_data function
-    * Initially, I read in roads and split by fips code to separate into states. This made
-    it difficult to deduplicate roads, since roads that were duplicated across states
-    would be deleted for one state and not the other. To avoid this, I read in all roads
-    at the beginning, sorted by hierarchy (primary, secondary, other), processed/exploded
-    (made all multilinestrings into linestrings), and deduplicated > then clipped by state
-    polygons.
-    * Code for State_Fips version exists in separate script (can include if necessary).
-
-    """
-
-    for year in year_range:
-        state_region_join(year)
-
-    # Join Proportional and Proxy Data
-    ####################################################################################
-
-    # Join Proportional and Proxy Data
-    roads_proxy = prop_proxy_join()
-
-    # Save output
-    roads_proxy.to_parquet(reporting_proxy_output)
-    return None
-
-########################################################################################
-########################################################################################
-########################################################################################
-########################################################################################
-# %% Processed_state_roads method described above
-states_gdf = read_states()
-region_gdf = read_regions(2012)
-roads_2012 = gpd.read_parquet(task_outputs_path / "reduced_roads_2012.parquet")
-
-import time
-import geopandas as gpd
-import pandas as pd
-from tqdm import tqdm
-from datetime import datetime
-
-def process_roads(road_data, states_gdf):
-    start_time = time.time()
-    final_roads = []
-
-    # Process each state
-    for _, state in tqdm(states_gdf.iterrows(), desc="Processing states"):
-        # Create single state GeoDataFrame
-        print(f'Reading states gdf {datetime.now()}')
-        state_gdf = gpd.GeoDataFrame(geometry=[state.geometry], crs=states_gdf.crs)
-
-        # Clip roads to state boundary
-        state_roads = gpd.clip(road_data, state_gdf)
-
-        if not state_roads.empty:
-            # Add state identifier
-            state_roads['STUSPS'] = state.STUSPS
+        if not interm_out_path.exists():
+            # create an empty CSV file to store intermediate results
+            pd.DataFrame(columns=['state', 'region', 'year', 'road_type', 'cell_id', 'vehicle', 'proxy', 'rd_length']).to_csv(interm_out_path, index=False)
         
-        final_roads.append(state_roads)
-        print(f"Processed {state.STUSPS}")
+        else:
+            # read in the intermediate file and filter cells that have already been processed based on ['cell_id', 'state', 'region']
+            processed_cells = pd.read_csv(interm_out_path)
+            cells = cells.loc[~cells.set_index(['cell_id', 'state', 'region']).index.isin(processed_cells.set_index(['cell_id', 'state', 'region']).index)]
+            # remove the intermediate file from environment
+            del(processed_cells)
 
-    # Combine all results
-    final_result = pd.concat(final_roads)
+        cell_n = len(cells)
+        for idx, cell in cells.reset_index().iterrows():
+            print(f'Processing index {idx}/{cell_n}', end='\r')
+            cell_id = cell['cell_id']
+            state = cell['state']
+            region = cell['region']
+            geom = cell['geometry'] # get box around cell for more efficient spatial query
 
-    end_time = time.time()
-    execution_time = end_time - start_time
-    print(f"Total execution time: {execution_time:.2f} seconds")
+            # get road value from proxy data
+            road_value = road_proxy_df.loc[(road_proxy_df['state'] == state) & (road_proxy_df['region'] == region) & (road_proxy_df['year'] == year)]
 
-    return final_result
+            # get roads in cell
+            try:
+                cell_roads = intersect_and_clip(roads, geom)
 
-processed_roads_2012 = process_roads(roads_2012, states_gdf)
-# 2 hr 50 min
+            except se.GEOSException as e:
+                geom = geom.buffer(0)
 
-# Make all LineStrings
-processed_roads_2012 = process_geometry_column(processed_roads_2012)
-
-# Write out
-#processed_roads_2012.to_parquet(task_outputs_path / "processed_state_roads_2012.parquet")
-processed_roads_2012 = gpd.read_parquet(task_outputs_path / "processed_state_roads_2012.parquet")
-
-########################################################################################
-########################################################################################
-# %% Dask (commented out sections during testing) and Not Dask: Previous Method for state joins
-
-import dask_geopandas as dgpd
-
-"""
-Dask code is messy because I ran into issues with deduplication, and attempted to
-troubleshoot everything in Dask.
-In hindsight, I should have just used my reduce method applied above and attempted
-dask on the region split and/or the state join.
-"""
-
-year_range = [2020, 2021, 2022]
-
-
-def read_reduce_data(year):
-    """
-    Read in All Roads data. Deduplicate and reduce data early.
-    """
-    road_data = (gpd.read_parquet(f"{raw_road_file}{year}_us_allroads.parquet",
-                                  columns=['MTFCC', 'geometry'])
-                 .assign(year=year))
-
-    print(type(road_data))
-
-    road_data = dgpd.from_geopandas(road_data, npartitions=4)
-
-    print("Type after Dask transformation:", type(road_data))
-    print("Columns:", road_data.columns)
-
-    # Order road types
-    road_type_order = ['Primary', 'Secondary', 'Other']
-
-    # Define road type
-    # def assign_road_type(df):
-    #     df['road_type'] = (np.select(
-    #         [
-    #             df['MTFCC'] == 'S1100',
-    #             df['MTFCC'] == 'S1200',
-    #             df['MTFCC'].isin(['S1400', 'S1630', 'S1640'])
-    #         ],
-    #         [
-    #             'Primary',
-    #             'Secondary',
-    #             'Other'
-    #         ],
-    #         default=None
-    #     )
-    #     return df
-
-    road_data = (
-        road_data.map_partitions(
-            lambda df: df.assign(
-                road_type=pd.Categorical(
-                    np.select(
-                        [
-                            df['MTFCC'] == 'S1100',
-                            df['MTFCC'] == 'S1200',
-                            df['MTFCC'].isin(['S1400', 'S1630', 'S1640'])
-                        ],
-                        [
-                            'Primary',
-                            'Secondary',
-                            'Other'
-                        ],
-                        default=None
-                    ),
-                    categories=road_type_order,
-                    ordered=True
+                try:
+                    cell_roads = intersect_and_clip(roads, geom)
+            
+                except Exception as e:
+                    print(f'ERROR processing cell {cell_id}, region: {region}, state: {state}, for year {year}: {e}')
+                    continue
+            # Calculate the length of lines within each road_type group
+            # first check if DF has values. intersect_and_clip returns an empty DF if no roads are found
+            if not cell_roads.empty:
+                road_type_df = (
+                    cell_roads
+                    .groupby(road_type_col, observed=True)['geometry']
+                    .apply(lambda x: x.length.sum()).rename('rd_length')
+                    .to_frame().reset_index()
+                    .assign(state=state, region=region, year=year, cell_id=cell_id)
+                    .set_index(['state', 'region', 'year', 'cell_id', road_type_col])
                 )
-            ),
-            meta={'MTFCC': 'object',
-                  'geometry': 'geometry',
-                  'year': 'int64',
-                  'road_type': 'category'}
-        )
-    )
+            else:
+                # if no roads in cell, create a DF where rd_length are 0 and join
+                road_type_df = pd.DataFrame({road_type_col: roads[road_type_col].unique(), 'rd_length': 0}).assign(state=state, region=region, year=year, cell_id=cell_id, rd_length=0).set_index(['state', 'region', 'year', 'cell_id', road_type_col])
 
-    print("Columns after road_type_creation:", road_data.columns)
-    print("Type after road_type_creation:", type(road_data))
+            # merge with road_value and write to CSV
+            road_value.set_index(['state', 'region', 'year', 'road_type']).join(road_type_df, how='left').reset_index()\
+                .to_csv(interm_out_path, mode='a', header=False, index=False)
 
-    # Convert back to Dask GeoDataFrame
-    def dask_to_dgdf(dask_df):
-        return dgpd.from_dask_dataframe(dask_df, geometry='geometry')
+        # concatenate
+        long_table = pd.read_csv(interm_out_path)
 
-    # Convert back
-    road_data = dask_to_dgdf(road_data)
-    road_data = road_data.set_crs("ESRI:102003", allow_override=True)
+        print(f'\nGridwise processing for {year} complete: {datetime.now()}')
 
-    #road_data = road_data[['year', 'road_type', 'geometry']]
-    #print("Columns after column reduction:", road_data.columns)
+        print(f'Calculating emission allocatoin for {year}')
 
-    # Order road types
-    # road_type_order = ['Primary', 'Secondary', 'Other']
+        # calculate road length by state:
+        state_road_length = long_table.groupby(['state', 'year'])['rd_length'].sum().rename('m_road_in_state').reset_index()
 
-    #road_data['road_type'] = pd.Categorical(road_data['road_type'], categories=road_type_order, ordered=True)
-    #road_data = road_data.sort_values('road_type').reset_index(drop=True)
-    def sort_by_road_type(road_data):
-        road_data_sorted = road_data.map_partitions(
-            lambda df: df.sort_values('road_type').reset_index(drop=True),
-            meta={'MTFCC': 'object',
-                  'geometry': 'geometry',
-                  'year': 'int64',
-                  'road_type': 'category'}
-        )
-        return road_data_sorted
+        # join with long table:
+        long_table = long_table.join(state_road_length.set_index(['state', 'year']), on=['state', 'year'])
 
-    road_data = sort_by_road_type(road_data)
+        # calculate methane emission allocation at the cell/region level
+        long_table['methane_emission_allocation'] = (long_table['rd_length'] / long_table['m_road_in_state']) * long_table['proxy']
+ 
+        print(f'Saving to CSV... {datetime.now()}')
+        long_table.to_csv(task_outputs_path / f"gridwise_mapping_{year}.csv", index=False)
 
-    print("Columns after sort:", road_data.columns)
+        print(f'FULL PROCESSING FOR {year} COMPLETE: {datetime.now()}')
 
-    print("Before explode:", type(road_data))
-    # Convert back
-    road_data = dask_to_dgdf(road_data)
-    road_data = road_data.set_crs("ESRI:102003", allow_override=True)
-    print("After convert back:", type(road_data))
-
-    # Explode to make LineStrings
-    road_data = road_data.map_partitions(
-        lambda df: gpd.GeoDataFrame(df).explode(column='geometry')
-    ).reset_index(drop=True)
-
-    print("After explode:", type(road_data))
-    road_data = dask_to_dgdf(road_data)
-    road_data = road_data.set_crs("ESRI:102003", allow_override=True)
-    print("After convert back:", type(road_data))
-    # Remove duplicates of geometries
-    road_data = road_data.drop_duplicates(subset='geometry')
-
-    print("After drop_duplicates:", type(road_data))
-    road_data = dask_to_dgdf(road_data)
-    road_data = road_data.set_crs("ESRI:102003", allow_override=True)
-
-    # Compute
-    road_data = road_data.compute()
-
-    print("After compute", type(road_data))
-
-    if not isinstance(road_data, gpd.GeoDataFrame):
-        road_data = gpd.GeoDataFrame(road_data, crs="ESRI:102003")
-
-    print(type(road_data))
-
-    # Adjust columns
-    road_data = road_data[['year', 'road_type', 'geometry']]
-
-    road_data = dgpd.from_geopandas(road_data, npartitions=4)
-
-    print("type after transform:", type(road_data))
-
-    # Separate out Road Types
-    prim_year = road_data[road_data['road_type'] == 'Primary']
-    sec_year = road_data[road_data['road_type'] == 'Secondary']
-    oth_year = road_data[road_data['road_type'] == 'Other']
-
-    buffer_distance = 3  # meters
-
-    meta = {
-        'year': 'int64',
-        'road_type': 'category',
-        'geometry': 'geometry'
-    }
-
-    # Define buffer_geometries
-    def buffer_geometries(gdf, buffer_distance):
-        # Buffer geometries
-        gdf['geometry'] = gdf.geometry.buffer(buffer_distance)
-        return gdf
-
-    def apply_overlay(target_gdf, buffer_gdf):
-        return gpd.overlay(target_gdf.compute(), buffer_gdf.compute(), how='difference')
-
-    def process_year_data(prim_year, sec_year, oth_year, buffer_distance):
-        # Set buffers
-        prim_year_buffer = prim_year.map_partitions(buffer_geometries, buffer_distance, meta=meta)
-        prisec_year = dd.concat([prim_year, sec_year], axis=0)
-        prisec_buffer = prisec_year.map_partitions(buffer_geometries, buffer_distance, meta=meta)
-
-        # Apply overlay
-        sec_red = apply_overlay(sec_year, prim_year_buffer)
-        other_red = apply_overlay(oth_year, prisec_buffer)
-
-        return prim_year, sec_red, other_red
-
-    prim_year, sec_red, other_red = process_year_data(prim_year, sec_year, oth_year, buffer_distance)
-
-    # Combine
-    road_data = dd.concat([prim_year, sec_red, other_red], axis=0)
-
-    road_data = road_data[['year', 'road_type', 'geometry']]
-
-    # Compute
-    road_data = road_data.compute()
-
-    road_data.to_parquet(task_outputs_path / f"reduced_roads_{year}.parquet")
-
-    del road_data, prim_year, sec_year, oth_year
-
-    gc.collect()
-
-    return None
+        del(roads)
 
 
-######
-# my_road_data = read_reduce_data()
-for year in year_range:
-    read_reduce_data(year)
-    print(f"Reduced roads for {year}")
-######
 
+        # The result of the above output is a CSV file with the following columns:
+        # - state
+        # - region
+        # - year
+        # - road_type
+        # - proxy
+        # - rd_length
+        # - cell_id
+        # - m_road_in_state 
+        # - methane_emission
 
-########################################################################################
-# %% Non Dask Version
+        # NEXT AGGREGATING TO THE CELL/YEAR level
+        # The variable for methane emission allocation represents the proportion of 
+        # state-wide methane emissions from that year that are attirbutable to that vehicle type/road/urbanicity(region) type in a given cell
+        long_table.groupby(['year', 'cell_id'])['methane_emission_allocation'].sum()\
+            .reset_index().to_csv(task_outputs_path / f"gridwise_mapping_{year}_cell.csv", index=False)
 
-year_range = [2020, 2021]
-
-
-def read_reduce_data_nodask(year):
-    """
-    Read in All Roads data. Deduplicate and reduce data early.
-    """
-    # Order road types
-    road_type_order = ['Primary', 'Secondary', 'Other']
-
-    df = (gpd.read_parquet(f"{raw_road_file}{year}_us_allroads.parquet",
-                           columns=['MTFCC', 'geometry'])
-          .assign(year=year))
-
-    road_data = (
-        df.to_crs("ESRI:102003")
-        .assign(
-            geometry=lambda df: df.normalize(),
-            road_type=lambda df: pd.Categorical(
-                np.select(
-                    [
-                        df['MTFCC'] == 'S1100',
-                        df['MTFCC'] == 'S1200',
-                        df['MTFCC'].isin(['S1400', 'S1630', 'S1640'])
-                    ],
-                    [
-                        'Primary',
-                        'Secondary',
-                        'Other'
-                    ],
-                    default=None
-                ),
-                categories=road_type_order,  # Define the categories
-                ordered=True  # Ensure the categories are ordered
-            )
-        )
-    )
-    # Sort
-    road_data = road_data.sort_values('road_type').reset_index(drop=True)
-    # Explode to make LineStrings
-    road_data = road_data.explode(index_parts=True).reset_index(drop=True)
-    # Remove duplicates of geometries
-    road_data = road_data.drop_duplicates(subset='geometry', keep='first')
-
-    # Separate out Road Types
-    prim_year = road_data[road_data['road_type'] == 'Primary']
-    sec_year = road_data[road_data['road_type'] == 'Secondary']
-    oth_year = road_data[road_data['road_type'] == 'Other']
-
-    buffer_distance = 3  # meters
-
-    # Set buffers
-    prim_buffer = prim_year.buffer(buffer_distance)
-    prim_buffer = gpd.GeoDataFrame(geometry=prim_buffer, crs=road_data.crs)
-
-    prisec_buffer = pd.concat([prim_year, sec_year], ignore_index=True)
-    prisec_buffer = prisec_buffer.buffer(buffer_distance)
-    prisec_buffer = gpd.GeoDataFrame(geometry=prisec_buffer, crs=road_data.crs)
-
-    # Overlay
-    sec_red = gpd.overlay(sec_year, prim_buffer, how='difference')
-    other_red = gpd.overlay(oth_year, prisec_buffer, how='difference')
-
-    # Combine
-    road_data = pd.concat([prim_year, sec_red, other_red], ignore_index=True)
-
-    road_data = road_data[['year', 'road_type', 'geometry']]
-
-    # Write to parquet
-    road_data.to_parquet(task_outputs_path / f"reduced_roads_{year}.parquet")
-
-    del road_data, prim_year, sec_year, oth_year, prim_buffer, prisec_buffer, sec_red, other_red
-
-    gc.collect()
-
-    return None
-
-########################################################################################
-
-
-for year in year_range:
-    read_reduce_data_nodask(year)
-    print(f"Reduced roads for {year}")
